clinical-interface/backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1
databases==0.8.0
aiosqlite==0.19.0

# Async Task Queue
celery==5.3.4
redis==5.0.1

# Medical Imaging
nibabel==5.2.0
SimpleITK==2.3.1
zarr==2.16.1
numcodecs==0.12.1

# Deep Learning (inherit from main requirements)
torch>=2.0.0
torchvision>=0.15.0
monai==1.3.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0

# Report Generation
reportlab==4.0.7
matplotlib==3.8.2
pillow==10.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
